requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery==5.3.4
celery[redis]==5.3.4

# GitHub Integration
PyGithub==1.59.1
cryptography==41.0.8

# LLM Integration
openai==1.3.7

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Environment & Config
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
